build: Support sysprof-6
Sysprof 45 changed the API version of the main library. We mainly
depend on sysprof-capture which is *not* affected, except for
looking up the datadir where sysprof's D-Bus interfaces are
installed.

That lookup doesn't care whether we are using sysprof-6 or
sysprof-4, so accept either dependency.

https://gitlab.gnome.org/GNOME/mutter/-/issues/2962

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3187>