backends/native: Prevent crash on resume
When first resuming from a suspend, the renderer
views may not yet have realized cursor stage views.

That can lead to a crash as the monitors light up
and force a cursor update leading to invalidation
of an unallocated cursor stage view.

This commit avoids the crash by doing a NULL check.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3012