util: Add way to print backtraces
Now that mutter can use a realtime thread its very important that
it doesn't stall for too long, since that can result in the kernel
killing it.  Ironically, a main reason mutter could stall is
kernel bugs.

When a stall happens, we need a way to see why. This commit adds
a new function, "meta_print_backtrace" to print backtrace of the
current process and the kernel (if possible).

A future commit will use this new function.